requirements.txt
-f /usr/share/pip-wheels
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
bcolz==1.0.0
Bottleneck==1.0.0
certifi==2016.2.28
click==6.6
configparser==3.3.0.post2
contextlib2==0.5.1
cycler==0.10.0
cyordereddict==1.0.0
Cython==0.24
decorator==4.0.9
dill==0.2.5
Django==1.9.5
django-cleanup==0.4.2
entrypoints==0.2
functools32==3.2.3.post2
gunicorn==19.4.5
ipykernel==4.3.1
ipython==4.2.0
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.2.2
jupyter-core==4.1.0
Lasagne==0.1
Logbook==0.12.5
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.7.2
nbconvert==4.2.0
nbformat==4.0.1
networkx==1.11
notebook==4.2.0
numexpr==2.5.2
numpy==1.11.0
pandas==0.18.0
pathlib2==2.1.0
patsy==0.4.1
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
ptyprocess==0.5.1
Pygments==2.1.3
pyparsing==2.1.1
PySide==1.2.4
python-dateutil==2.5.3
pytz==2016.4
pyzmq==15.2.0
qtconsole==4.2.1
requests==2.9.1
scikit-learn==0.17.1
scipy==0.17.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
SQLAlchemy==1.0.12
statsmodels==0.6.1
terminado==0.6
Theano==0.8.2
toolz==0.7.4
tornado==4.3
traitlets==4.2.1
zipline==0.9.0
